Remove the usage of the label trait so that multiple labels can be attributed to one test
If multiple traits of name "Label" (but different values) are provided
in the traits collection, VS will display only one of them.  (most
probably it is generating an unique collection).
Code is changed so that the label name is what was before the label
value. The new label value is an empty string.